feat: add possibility to configure http client timeout and health checker interval
There are new fields in the issuer's kinds to allow
defining the values of newly added configuration
options. Additionally, to better group the fields
responsible for various funcionalitites, a refactor
was made. The previous exisiting option, whose
way of definition has changed, can still be given
in the old way but the newest is prefered.